+++ ostap/stats/tests/test_stats_2samples.py
@@ -9,7 +9,7 @@
 """
 # ==============================================================================
 from   ostap.stats.twosamples import TSTest, TSToys 
-from   ostap.core.meta_info   import python_info 
+from   ostap.core.meta_info   import python_info, root_info  
 from   ostap.plotting.canvas  import use_canvas
 from   ostap.logger.pretty    import pretty_float
 from   ostap.utils.utils      import vrange 
@@ -21,19 +21,22 @@ from   ostap.logger.logger import getLogger
 if '__main__' ==  __name__ : logger = getLogger ( 'tests_stats_2samples' )
 else                       : logger = getLogger ( __name__            )
 # ==============================================================================
-N1  = 200
-N2  = 200
+N1  = 150
+N2  = 150
 
 ds1 = array.array ( 'd' , ( random.gauss ( 0   , 1   ) for i in range ( N1 ) ) ) ## Gaussian 
 ds2 = array.array ( 'd' , ( random.gauss ( 0   , 1   ) for i in range ( N2 ) ) ) ## the same  Gaussian 
 ds3 = array.array ( 'd' , ( random.gauss ( 0.5 , 1   ) for i in range ( N2 ) ) ) ## different Gaussian 
-ds4 = array.array ( 'd' , ( random.gauss ( 0   , 1.3 ) for i in range ( N2 ) ) ) ## different Gaussian 
+ds4 = array.array ( 'd' , ( random.gauss ( 0   , 1.5 ) for i in range ( N2 ) ) ) ## different Gaussian 
 
 # =============================================================================
 def test_2samples_good () :
     
     logger = getLogger ( 'test_2samples_good' ) 
-
+    if root_info < ( 6, 24 ) :
+        logger.warning ( 'Test is disabled' )
+        return
+    
     test   = TSTest ( ds1 , ds2 )
     title  = 'Test good samples'
     logger.info ( '%s:\n%s' % ( title , test.table ( title = title , prefix = '# ' ) ) )
@@ -42,15 +45,23 @@ def test_2samples_good () :
     title  = 'Toys good samples'
     logger.info ( '%s:\n%s' % ( title , toys.table ( title = title , prefix = '# ' ) ) )
 
-    with use_canvas ( 'test_2samples_good/test' , wait = 3 ) : test.draw() 
-    with use_canvas ( 'test_2samples_good/toys' , wait = 3 ) : toys.draw() 
-
+    with use_canvas ( 'test_2samples_good/test'                    , wait = 1 ) : test.draw()    
+    with use_canvas ( 'test_2samples_good/toys Kolmogorov-Smirnov' , wait = 1 ) : toys.draw ('KS') 
+    with use_canvas ( 'test_2samples_good/toys Kuiper'             , wait = 1 ) : toys.draw ('K' ) 
+    with use_canvas ( 'test_2samples_good/toys Anderson-Darling'   , wait = 1 ) : toys.draw ('AD') 
+    with use_canvas ( 'test_2samples_good/toys Cramers-von Mises'  , wait = 1 ) : toys.draw ('CM') 
+    with use_canvas ( 'test_2samples_good/toys Zhang ZK'           , wait = 1 ) : toys.draw ('ZK') 
+    with use_canvas ( 'test_2samples_good/toys Zhang ZA'           , wait = 1 ) : toys.draw ('ZA') 
+    with use_canvas ( 'test_2samples_good/toys Zhang ZC'           , wait = 1 ) : toys.draw ('ZC') 
     
 # =============================================================================
 def test_2samples_same () :
     
     logger = getLogger ( 'test_2samples_same' ) 
-    
+    if root_info < ( 6, 24 ) :
+        logger.warning ( 'Test is disabled' )
+        return
+        
     test   = TSTest ( ds1 , ds1 )                   
     title  = 'Test same samples'
     logger.info ( '%s:\n%s' % ( title , test.table ( title = title , prefix = '# ' ) ) )
@@ -59,13 +70,22 @@ def test_2samples_same () :
     title  = 'Toys same samples'
     logger.info ( '%s:\n%s' % ( title , toys.table ( title = title , prefix = '# ' ) ) )
 
-    with use_canvas ( 'test_2samples_same/test' , wait = 3 ) : test.draw() 
-    with use_canvas ( 'test_2samples_same/toys' , wait = 3 ) : toys.draw() 
+    with use_canvas ( 'test_2samples_same/test'                    , wait = 1 ) : test.draw()    
+    with use_canvas ( 'test_2samples_same/toys Kolmogorov-Smirnov' , wait = 1 ) : toys.draw ('KS') 
+    with use_canvas ( 'test_2samples_same/toys Kuiper'             , wait = 1 ) : toys.draw ('K' ) 
+    with use_canvas ( 'test_2samples_same/toys Anderson-Darling'   , wait = 1 ) : toys.draw ('AD') 
+    with use_canvas ( 'test_2samples_same/toys Cramers-von Mises'  , wait = 1 ) : toys.draw ('CM') 
+    with use_canvas ( 'test_2samples_same/toys Zhang ZK'           , wait = 1 ) : toys.draw ('ZK') 
+    with use_canvas ( 'test_2samples_same/toys Zhang ZA'           , wait = 1 ) : toys.draw ('ZA') 
+    with use_canvas ( 'test_2samples_same/toys Zhang ZC'           , wait = 1 ) : toys.draw ('ZC') 
     
 # =============================================================================
 def test_2samples_bad1 () :
     
     logger = getLogger ( 'test_2samples_bad1'  ) 
+    if root_info < ( 6, 24 ) :
+        logger.warning ( 'Test is disabled' )
+        return
     
     test   = TSTest ( ds1 , ds3 )
     title  = 'Test bad/1 samples'
@@ -75,13 +95,22 @@ def test_2samples_bad1 () :
     title  = 'Toys bad/1 samples'
     logger.info ( '%s:\n%s' % ( title , toys.table ( title = title , prefix = '# ' ) ) )
     
-    with use_canvas ( 'test_2samples_bad1/test' , wait = 3 ) : test.draw() 
-    with use_canvas ( 'test_2samples_bad1/toys' , wait = 3 ) : toys.draw() 
+    with use_canvas ( 'test_2samples_bad1/test'                    , wait = 1 ) : test.draw()    
+    with use_canvas ( 'test_2samples_bad1/toys Kolmogorov-Smirnov' , wait = 1 ) : toys.draw ('KS') 
+    with use_canvas ( 'test_2samples_bad1/toys Kuiper'             , wait = 1 ) : toys.draw ('K' ) 
+    with use_canvas ( 'test_2samples_bad1/toys Anderson-Darling'   , wait = 1 ) : toys.draw ('AD') 
+    with use_canvas ( 'test_2samples_bad1/toys Cramers-von Mises'  , wait = 1 ) : toys.draw ('CM') 
+    with use_canvas ( 'test_2samples_bad1/toys Zhang ZK'           , wait = 1 ) : toys.draw ('ZK') 
+    with use_canvas ( 'test_2samples_bad1/toys Zhang ZA'           , wait = 1 ) : toys.draw ('ZA') 
+    with use_canvas ( 'test_2samples_bad1/toys Zhang ZC'           , wait = 1 ) : toys.draw ('ZC') 
     
 # =============================================================================
 def test_2samples_bad2 () :
     
     logger = getLogger ( 'test_2samples_bad2'  ) 
+    if root_info < ( 6, 24 ) :
+        logger.warning ( 'Test is disabled' )
+        return
     
     test   = TSTest ( ds1 , ds4  )
     title  = 'Test bad/2 samples'
@@ -91,8 +120,14 @@ def test_2samples_bad2 () :
     title  = 'Toys bad/2 samples'
     logger.info ( '%s:\n%s' % ( title , toys.table ( title = title , prefix = '# ' ) ) )
     
-    with use_canvas ( 'test_2samples_bad2/test' , wait = 3 ) : test.draw() 
-    with use_canvas ( 'test_2samples_bad2/toys' , wait = 3 ) : toys.draw() 
+    with use_canvas ( 'test_2samples_bad2/test'                    , wait = 1 ) : test.draw()    
+    with use_canvas ( 'test_2samples_bad2/toys Kolmogorov-Smirnov' , wait = 1 ) : toys.draw ('KS') 
+    with use_canvas ( 'test_2samples_bad2/toys Kuiper'             , wait = 1 ) : toys.draw ('K' ) 
+    with use_canvas ( 'test_2samples_bad2/toys Anderson-Darling'   , wait = 1 ) : toys.draw ('AD') 
+    with use_canvas ( 'test_2samples_bad2/toys Cramers-von Mises'  , wait = 1 ) : toys.draw ('CM') 
+    with use_canvas ( 'test_2samples_bad2/toys Zhang ZK'           , wait = 1 ) : toys.draw ('ZK') 
+    with use_canvas ( 'test_2samples_bad2/toys Zhang ZA'           , wait = 1 ) : toys.draw ('ZA') 
+    with use_canvas ( 'test_2samples_bad2/toys Zhang ZC'           , wait = 1 ) : toys.draw ('ZC') 
 
 # ===============================================================================
 if '__main__' == __name__ :